+++ CHANGES.md
@@ -1,8 +1,14 @@
 ## Version 9.3.0 [NOT YET RELEASED]
 
+New languages:
+
+- *Tagger Script* by [Philipp Wolfer][]
+
 Improvements to existing languages and styles:
 
 - Improve GAMS language parsing
+
+[Philipp Wolfer]: https://github.com/phw
 
 ## Version 9.2.0
 
@@ -36,7 +42,6 @@ Improvements to existing languages and styles:
 [Michael Johnston]: https://github.com/lastobelus
 [Taras]: https://github.com/oxdef
 [Robert Dodier]: https://github.com/robert-dodier
-
 
 ## Version 9.1.0